+++ src/graphics/canvas.rs
@@ -1,5 +1,6 @@
 use std::rc::Rc;
 
+use crate::error::Result;
 use crate::glm::{self, Mat4};
 use crate::graphics::opengl::{GLDevice, GLFramebuffer};
 use crate::graphics::{DrawParams, Drawable, FilterMode, Texture};
@@ -23,7 +24,8 @@ pub struct Canvas {
 impl Canvas {
     /// Creates a new canvas.
     pub fn new(ctx: &mut Context, width: i32, height: i32) -> Canvas {
-        Canvas::with_device(&mut ctx.gl, width, height, true)
+        // TODO: Make this return Result in 0.3
+        Canvas::with_device(&mut ctx.gl, width, height, true).expect("Could not create canvas")
     }
 
     pub(crate) fn with_device(
@@ -31,20 +33,21 @@ impl Canvas {
         width: i32,
         height: i32,
         rebind_previous: bool,
-    ) -> Canvas {
-        let texture = Texture::with_device_empty(device, width, height);
-        let framebuffer = device.new_framebuffer();
+    ) -> Result<Canvas> {
+        let texture = Texture::with_device_empty(device, width, height)?;
+        let framebuffer = device.new_framebuffer()?;
+
         device.attach_texture_to_framebuffer(
             &framebuffer,
             &texture.handle.borrow(),
             rebind_previous,
         );
 
-        Canvas {
+        Ok(Canvas {
             texture,
             framebuffer: Rc::new(framebuffer),
             projection: glm::ortho(0.0, width as f32, 0.0, height as f32, -1.0, 1.0),
-        }
+        })
     }
 
     /// Returns the width of the canvas.