+++ src/graphics/opengl.rs
@@ -1,88 +1,86 @@
-use std::ffi::{CStr, CString};
 use std::mem;
-use std::ptr;
+use std::rc::Rc;
 
-use gl::{self, types::*};
+use glow::Context as _;
 
 use crate::error::{Result, TetraError};
 use crate::glm::Mat4;
 use crate::graphics::FilterMode;
 
+type GlContext = glow::native::Context;
+
+type BufferId = <GlContext as glow::Context>::Buffer;
+type ProgramId = <GlContext as glow::Context>::Program;
+type TextureId = <GlContext as glow::Context>::Texture;
+type FramebufferId = <GlContext as glow::Context>::Framebuffer;
+type VertexArrayId = <GlContext as glow::Context>::VertexArray;
+
 pub struct GLDevice {
-    current_vertex_buffer: GLuint,
-    current_index_buffer: GLuint,
-    current_program: GLuint,
-    current_texture: GLuint,
-    current_framebuffer: GLuint,
-    current_vertex_array: GLuint,
+    gl: Rc<GlContext>,
+
+    current_vertex_buffer: Option<BufferId>,
+    current_index_buffer: Option<BufferId>,
+    current_program: Option<ProgramId>,
+    current_texture: Option<TextureId>,
+    current_framebuffer: Option<FramebufferId>,
+    current_vertex_array: Option<VertexArrayId>,
 
     // TODO: I kinda don't like this being here, should probably be on the graphics context.
     default_filter_mode: FilterMode,
 }
 
 impl GLDevice {
-    pub fn new() -> GLDevice {
-        let mut current_vertex_array = 0;
-
+    pub fn new(gl: GlContext) -> Result<GLDevice> {
         unsafe {
-            gl::Enable(gl::CULL_FACE);
-            gl::Enable(gl::BLEND);
+            gl.enable(glow::CULL_FACE);
+            gl.enable(glow::BLEND);
 
             // This default might want to change if we introduce
             // custom blending modes.
-            gl::BlendFuncSeparate(
-                gl::SRC_ALPHA,
-                gl::ONE_MINUS_SRC_ALPHA,
-                gl::ONE,
-                gl::ONE_MINUS_SRC_ALPHA,
+            gl.blend_func_separate(
+                glow::SRC_ALPHA,
+                glow::ONE_MINUS_SRC_ALPHA,
+                glow::ONE,
+                glow::ONE_MINUS_SRC_ALPHA,
             );
 
             // This is only needed for Core GL - if we wanted to be uber compatible, we'd
             // turn it off on older versions.
-            gl::GenVertexArrays(1, &mut current_vertex_array);
-            gl::BindVertexArray(current_vertex_array);
+            let current_vertex_array = gl.create_vertex_array().map_err(TetraError::OpenGl)?;
+            gl.bind_vertex_array(Some(current_vertex_array));
 
-            println!(
-                "OpenGL Device: {}",
-                CStr::from_ptr(gl::GetString(gl::RENDERER) as *const _).to_string_lossy()
-            );
+            println!("OpenGL Device: {}", gl.get_parameter_string(glow::RENDERER));
+            println!("OpenGL Driver: {}", gl.get_parameter_string(glow::VERSION));
+            println!("OpenGL Vendor: {}", gl.get_parameter_string(glow::VENDOR));
 
-            println!(
-                "OpenGL Driver: {}",
-                CStr::from_ptr(gl::GetString(gl::VERSION) as *const _).to_string_lossy()
-            );
+            // TODO: Find a nice way of exposing this via the platform layer
+            // println!("Swap Interval: {:?}", video.gl_get_swap_interval());
 
-            println!(
-                "OpenGL Vendor: {}",
-                CStr::from_ptr(gl::GetString(gl::VENDOR) as *const _).to_string_lossy()
-            );
-        }
+            Ok(GLDevice {
+                gl: Rc::new(gl),
 
-        // TODO: Find a nice way of exposing this via the platform layer
-        // println!("Swap Interval: {:?}", video.gl_get_swap_interval());
+                current_vertex_buffer: None,
+                current_index_buffer: None,
+                current_program: None,
+                current_texture: None,
+                current_framebuffer: None,
+                current_vertex_array: Some(current_vertex_array),
 
-        GLDevice {
-            current_vertex_buffer: 0,
-            current_index_buffer: 0,
-            current_program: 0,
-            current_texture: 0,
-            current_framebuffer: 0,
-            current_vertex_array,
-
-            default_filter_mode: FilterMode::Nearest,
+                default_filter_mode: FilterMode::Nearest,
+            })
         }
     }
 
     pub fn clear(&mut self, r: f32, g: f32, b: f32, a: f32) {
         unsafe {
-            gl::ClearColor(r, g, b, a);
-            gl::Clear(gl::COLOR_BUFFER_BIT);
+            self.gl.clear_color(r, g, b, a);
+            self.gl.clear(glow::COLOR_BUFFER_BIT);
         }
     }
 
     pub fn front_face(&mut self, front_face: FrontFace) {
         unsafe {
-            gl::FrontFace(front_face.into());
+            self.gl.front_face(front_face.into());
         }
     }
 
@@ -91,23 +89,26 @@ impl GLDevice {
         count: usize,
         stride: usize,
         usage: BufferUsage,
-    ) -> GLVertexBuffer {
+    ) -> Result<GLVertexBuffer> {
         unsafe {
-            let mut id = 0;
-            gl::GenBuffers(1, &mut id);
+            let id = self.gl.create_buffer().map_err(TetraError::OpenGl)?;
 
-            let buffer = GLVertexBuffer { id, count, stride };
+            let buffer = GLVertexBuffer {
+                gl: Rc::clone(&self.gl),
+                id,
+                count,
+                stride,
+            };
 
-            self.bind_vertex_buffer(&buffer);
+            self.bind_vertex_buffer(Some(&buffer));
 
-            gl::BufferData(
-                gl::ARRAY_BUFFER,
-                (count * mem::size_of::<GLfloat>()) as GLsizeiptr,
-                ptr::null() as *const GLvoid,
+            self.gl.buffer_data_size(
+                glow::ARRAY_BUFFER,
+                (count * mem::size_of::<f32>()) as i32,
                 usage.into(),
             );
 
-            buffer
+            Ok(buffer)
         }
     }
 
@@ -121,81 +122,75 @@ impl GLDevice {
         // TODO: This feels a bit unergonomic...
 
         unsafe {
-            self.bind_vertex_buffer(buffer);
+            self.bind_vertex_buffer(Some(buffer));
 
-            gl::VertexAttribPointer(
+            self.gl.vertex_attrib_pointer_f32(
                 index,
                 size,
-                gl::FLOAT,
-                gl::FALSE,
-                (buffer.stride * mem::size_of::<GLfloat>()) as GLsizei,
-                (offset * mem::size_of::<GLfloat>()) as *const _,
+                glow::FLOAT,
+                false,
+                (buffer.stride * mem::size_of::<f32>()) as i32,
+                (offset * mem::size_of::<f32>()) as i32,
             );
 
-            gl::EnableVertexAttribArray(index);
+            self.gl.enable_vertex_attrib_array(index);
         }
     }
 
-    pub fn set_vertex_buffer_data(
-        &mut self,
-        buffer: &GLVertexBuffer,
-        data: &[GLfloat],
-        offset: usize,
-    ) {
+    pub fn set_vertex_buffer_data(&mut self, buffer: &GLVertexBuffer, data: &[f32], offset: usize) {
         unsafe {
-            assert!(offset + data.len() <= buffer.count);
-
-            self.bind_vertex_buffer(buffer);
+            self.bind_vertex_buffer(Some(buffer));
 
             // TODO: What if we want to discard what's already there?
+            // TODO: Is this cast safe?
 
-            gl::BufferSubData(
-                gl::ARRAY_BUFFER,
-                (offset * mem::size_of::<GLfloat>()) as GLsizeiptr,
-                (data.len() * mem::size_of::<GLfloat>()) as GLsizeiptr,
-                data.as_ptr() as *const GLvoid,
+            let byte_len = std::mem::size_of_val(data) / std::mem::size_of::<u8>();
+            let byte_slice = std::slice::from_raw_parts(data.as_ptr() as *const u8, byte_len);
+
+            self.gl.buffer_sub_data_u8_slice(
+                glow::ARRAY_BUFFER,
+                (offset * mem::size_of::<f32>()) as i32,
+                byte_slice,
             );
         }
     }
 
-    pub fn new_index_buffer(&mut self, count: usize, usage: BufferUsage) -> GLIndexBuffer {
+    pub fn new_index_buffer(&mut self, count: usize, usage: BufferUsage) -> Result<GLIndexBuffer> {
         unsafe {
-            let mut id = 0;
-            gl::GenBuffers(1, &mut id);
+            let id = self.gl.create_buffer().map_err(TetraError::OpenGl)?;
 
-            let buffer = GLIndexBuffer { id, count };
+            let buffer = GLIndexBuffer {
+                gl: Rc::clone(&self.gl),
+                id,
+                count,
+            };
 
-            self.bind_index_buffer(&buffer);
+            self.bind_index_buffer(Some(&buffer));
 
-            gl::BufferData(
-                gl::ELEMENT_ARRAY_BUFFER,
-                (count * mem::size_of::<GLuint>()) as GLsizeiptr,
-                ptr::null() as *const GLvoid,
+            self.gl.buffer_data_size(
+                glow::ELEMENT_ARRAY_BUFFER,
+                (count * mem::size_of::<u32>()) as i32,
                 usage.into(),
             );
 
-            buffer
+            Ok(buffer)
         }
     }
 
-    pub fn set_index_buffer_data(
-        &mut self,
-        buffer: &GLIndexBuffer,
-        data: &[GLuint],
-        offset: usize,
-    ) {
+    pub fn set_index_buffer_data(&mut self, buffer: &GLIndexBuffer, data: &[u32], offset: usize) {
         unsafe {
-            assert!(offset + data.len() <= buffer.count);
-
-            self.bind_index_buffer(buffer);
+            self.bind_index_buffer(Some(buffer));
 
             // TODO: What if we want to discard what's already there?
+            // TODO: Is this cast safe?
 
-            gl::BufferSubData(
-                gl::ELEMENT_ARRAY_BUFFER,
-                (offset * mem::size_of::<GLuint>()) as GLsizeiptr,
-                (data.len() * mem::size_of::<GLuint>()) as GLsizeiptr,
-                data.as_ptr() as *const GLvoid,
+            let byte_len = std::mem::size_of_val(data) / std::mem::size_of::<u8>();
+            let byte_slice = std::slice::from_raw_parts(data.as_ptr() as *const u8, byte_len);
+
+            self.gl.buffer_sub_data_u8_slice(
+                glow::ELEMENT_ARRAY_BUFFER,
+                (offset * mem::size_of::<u32>()) as i32,
+                byte_slice,
             );
         }
     }
@@ -206,113 +201,57 @@ impl GLDevice {
         fragment_shader: &str,
     ) -> Result<GLProgram> {
         unsafe {
-            let vertex_buffer = CString::new(vertex_shader).unwrap();
-            let fragment_buffer = CString::new(fragment_shader).unwrap();
-
-            let program_id = gl::CreateProgram();
+            let program_id = self.gl.create_program().map_err(TetraError::OpenGl)?;
 
             // TODO: IDK if this should be applied to *all* shaders...
-            let position_name = CString::new("a_position").unwrap();
-            let uv_name = CString::new("a_uv").unwrap();
-            let color_name = CString::new("a_color").unwrap();
-            let out_color_name = CString::new("o_color").unwrap();
+            self.gl.bind_attrib_location(program_id, 0, "a_position");
+            self.gl.bind_attrib_location(program_id, 1, "a_uv");
+            self.gl.bind_attrib_location(program_id, 2, "a_color");
+            self.gl.bind_frag_data_location(program_id, 0, "o_color");
 
-            gl::BindAttribLocation(program_id, 0, position_name.as_ptr());
-            gl::BindAttribLocation(program_id, 1, uv_name.as_ptr());
-            gl::BindAttribLocation(program_id, 2, color_name.as_ptr());
-            gl::BindFragDataLocation(program_id, 0, out_color_name.as_ptr());
+            let vertex_id = self
+                .gl
+                .create_shader(glow::VERTEX_SHADER)
+                .map_err(TetraError::OpenGl)?;
 
-            let vertex_id = gl::CreateShader(gl::VERTEX_SHADER);
-            gl::ShaderSource(vertex_id, 1, &vertex_buffer.as_ptr(), ptr::null());
-            gl::CompileShader(vertex_id);
-            gl::AttachShader(program_id, vertex_id);
+            self.gl.shader_source(vertex_id, vertex_shader);
+            self.gl.compile_shader(vertex_id);
+            self.gl.attach_shader(program_id, vertex_id);
 
-            let mut success = 0;
-            gl::GetShaderiv(vertex_id, gl::COMPILE_STATUS, &mut success);
-
-            if success != 1 {
-                return Err(TetraError::OpenGl(self.get_shader_info_log(vertex_id)));
+            if !self.gl.get_shader_compile_status(vertex_id) {
+                return Err(TetraError::OpenGl(self.gl.get_shader_info_log(vertex_id)));
             }
 
-            let fragment_id = gl::CreateShader(gl::FRAGMENT_SHADER);
-            gl::ShaderSource(fragment_id, 1, &fragment_buffer.as_ptr(), ptr::null());
-            gl::CompileShader(fragment_id);
-            gl::AttachShader(program_id, fragment_id);
+            let fragment_id = self
+                .gl
+                .create_shader(glow::FRAGMENT_SHADER)
+                .map_err(TetraError::OpenGl)?;
 
-            let mut success = 0;
-            gl::GetShaderiv(fragment_id, gl::COMPILE_STATUS, &mut success);
+            self.gl.shader_source(fragment_id, fragment_shader);
+            self.gl.compile_shader(fragment_id);
+            self.gl.attach_shader(program_id, fragment_id);
 
-            if success != 1 {
-                return Err(TetraError::OpenGl(self.get_shader_info_log(fragment_id)));
+            if !self.gl.get_shader_compile_status(vertex_id) {
+                return Err(TetraError::OpenGl(self.gl.get_shader_info_log(fragment_id)));
             }
 
-            gl::LinkProgram(program_id);
+            self.gl.link_program(program_id);
 
-            let mut success = 0;
-            gl::GetProgramiv(program_id, gl::LINK_STATUS, &mut success);
-
-            if success != 1 {
-                return Err(TetraError::OpenGl(self.get_program_info_log(program_id)));
+            if !self.gl.get_program_link_status(program_id) {
+                return Err(TetraError::OpenGl(self.gl.get_program_info_log(program_id)));
             }
 
-            gl::DeleteShader(vertex_id);
-            gl::DeleteShader(fragment_id);
+            self.gl.delete_shader(vertex_id);
+            self.gl.delete_shader(fragment_id);
 
-            let program = GLProgram { id: program_id };
+            let program = GLProgram {
+                gl: Rc::clone(&self.gl),
+                id: program_id,
+            };
 
             self.set_uniform(&program, "u_texture", 0);
 
             Ok(program)
-        }
-    }
-
-    fn get_shader_info_log(&mut self, shader_id: GLuint) -> String {
-        unsafe {
-            let mut max_len = 0;
-
-            gl::GetShaderiv(shader_id, gl::INFO_LOG_LENGTH, &mut max_len);
-
-            let mut result = vec![0_u8; max_len as usize];
-            let mut result_len = 0 as GLsizei;
-            gl::GetShaderInfoLog(
-                shader_id,
-                max_len as GLsizei,
-                &mut result_len,
-                result.as_mut_ptr() as *mut GLchar,
-            );
-
-            result.truncate(if result_len > 0 {
-                result_len as usize
-            } else {
-                0
-            });
-
-            String::from_utf8(result).unwrap()
-        }
-    }
-
-    fn get_program_info_log(&mut self, program_id: GLuint) -> String {
-        unsafe {
-            let mut max_len = 0;
-
-            gl::GetProgramiv(program_id, gl::INFO_LOG_LENGTH, &mut max_len);
-
-            let mut result = vec![0_u8; max_len as usize];
-            let mut result_len = 0 as GLsizei;
-            gl::GetProgramInfoLog(
-                program_id,
-                max_len as GLsizei,
-                &mut result_len,
-                result.as_mut_ptr() as *mut GLchar,
-            );
-
-            result.truncate(if result_len > 0 {
-                result_len as usize
-            } else {
-                0
-            });
-
-            String::from_utf8(result).unwrap()
         }
     }
 
@@ -321,61 +260,73 @@ impl GLDevice {
         T: UniformValue,
     {
         unsafe {
-            self.bind_program(program);
+            self.bind_program(Some(program));
 
-            let c_name = CString::new(name).unwrap();
-            let location = gl::GetUniformLocation(program.id, c_name.as_ptr());
-            value.set_uniform(location);
+            let location = self.gl.get_uniform_location(program.id, name);
+            value.set_uniform(program, location);
         }
     }
 
-    pub fn new_texture(&mut self, width: i32, height: i32, format: TextureFormat) -> GLTexture {
+    pub fn new_texture(
+        &mut self,
+        width: i32,
+        height: i32,
+        format: TextureFormat,
+    ) -> Result<GLTexture> {
         // TODO: I don't think we need mipmaps?
         unsafe {
-            let mut id = 0;
-            gl::GenTextures(1, &mut id);
+            let id = self.gl.create_texture().map_err(TetraError::OpenGl)?;
 
             let texture = GLTexture {
+                gl: Rc::clone(&self.gl),
+
                 id,
                 width,
                 height,
                 filter_mode: self.default_filter_mode,
             };
 
-            self.bind_texture(&texture);
+            self.bind_texture(Some(&texture));
 
-            let gl_filter_mode: GLenum = self.default_filter_mode.into();
+            self.gl
+                .tex_parameter_i32(glow::TEXTURE_2D, glow::TEXTURE_WRAP_S, glow::REPEAT as i32);
 
-            gl::TexParameteri(gl::TEXTURE_2D, gl::TEXTURE_WRAP_S, gl::REPEAT as GLint);
-            gl::TexParameteri(gl::TEXTURE_2D, gl::TEXTURE_WRAP_T, gl::REPEAT as GLint);
-            gl::TexParameteri(
-                gl::TEXTURE_2D,
-                gl::TEXTURE_MIN_FILTER,
-                gl_filter_mode as GLint,
+            self.gl
+                .tex_parameter_i32(glow::TEXTURE_2D, glow::TEXTURE_WRAP_T, glow::REPEAT as i32);
+
+            self.gl.tex_parameter_i32(
+                glow::TEXTURE_2D,
+                glow::TEXTURE_MIN_FILTER,
+                self.default_filter_mode.into(),
             );
-            gl::TexParameteri(
-                gl::TEXTURE_2D,
-                gl::TEXTURE_MAG_FILTER,
-                gl_filter_mode as GLint,
+
+            self.gl.tex_parameter_i32(
+                glow::TEXTURE_2D,
+                glow::TEXTURE_MAG_FILTER,
+                self.default_filter_mode.into(),
             );
-            gl::TexParameteri(gl::TEXTURE_2D, gl::TEXTURE_BASE_LEVEL, 0);
-            gl::TexParameteri(gl::TEXTURE_2D, gl::TEXTURE_MAX_LEVEL, 0);
+
+            self.gl
+                .tex_parameter_i32(glow::TEXTURE_2D, glow::TEXTURE_BASE_LEVEL, 0);
+
+            self.gl
+                .tex_parameter_i32(glow::TEXTURE_2D, glow::TEXTURE_MAX_LEVEL, 0);
 
             let format = format.into();
 
-            gl::TexImage2D(
-                gl::TEXTURE_2D,
+            self.gl.tex_image_2d(
+                glow::TEXTURE_2D,
                 0,
-                format as GLint, // love 2 deal with legacy apis
+                format as i32, // love 2 deal with legacy apis
                 width,
                 height,
                 0,
                 format,
-                gl::UNSIGNED_BYTE,
-                ptr::null(),
+                glow::UNSIGNED_BYTE,
+                None,
             );
 
-            texture
+            Ok(texture)
         }
     }
 
@@ -390,58 +341,54 @@ impl GLDevice {
         format: TextureFormat,
     ) {
         unsafe {
-            self.bind_texture(texture);
+            self.bind_texture(Some(texture));
 
-            let format = format.into();
-
-            gl::TexSubImage2D(
-                gl::TEXTURE_2D,
+            self.gl.tex_sub_image_2d_u8_slice(
+                glow::TEXTURE_2D,
                 0,
                 x,
                 y,
                 width,
                 height,
-                format,
-                gl::UNSIGNED_BYTE,
-                data.as_ptr() as *const GLvoid,
+                format.into(),
+                glow::UNSIGNED_BYTE,
+                Some(data),
             )
         }
     }
 
     pub fn set_texture_min_filter(&mut self, texture: &GLTexture, filter_mode: FilterMode) {
-        self.bind_texture(&texture);
-
-        let gl_filter_mode: GLenum = filter_mode.into();
+        self.bind_texture(Some(texture));
 
         unsafe {
-            gl::TexParameteri(
-                gl::TEXTURE_2D,
-                gl::TEXTURE_MIN_FILTER,
-                gl_filter_mode as GLint,
+            self.gl.tex_parameter_i32(
+                glow::TEXTURE_2D,
+                glow::TEXTURE_MIN_FILTER,
+                filter_mode.into(),
             );
         }
     }
 
     pub fn set_texture_mag_filter(&mut self, texture: &GLTexture, filter_mode: FilterMode) {
-        self.bind_texture(&texture);
-
-        let gl_filter_mode: GLenum = filter_mode.into();
+        self.bind_texture(Some(texture));
 
         unsafe {
-            gl::TexParameteri(
-                gl::TEXTURE_2D,
-                gl::TEXTURE_MAG_FILTER,
-                gl_filter_mode as GLint,
+            self.gl.tex_parameter_i32(
+                glow::TEXTURE_2D,
+                glow::TEXTURE_MAG_FILTER,
+                filter_mode.into(),
             );
         }
     }
 
-    pub fn new_framebuffer(&mut self) -> GLFramebuffer {
+    pub fn new_framebuffer(&mut self) -> Result<GLFramebuffer> {
         unsafe {
-            let mut id = 0;
-            gl::GenFramebuffers(1, &mut id);
+            let id = self.gl.create_framebuffer().map_err(TetraError::OpenGl)?;
 
-            GLFramebuffer { id }
+            Ok(GLFramebuffer {
+                gl: Rc::clone(&self.gl),
+                id,
+            })
         }
     }
 
@@ -454,18 +401,18 @@ impl GLDevice {
         unsafe {
             let previous_id = self.current_framebuffer;
 
-            self.bind_framebuffer(framebuffer);
+            self.bind_framebuffer(Some(framebuffer));
 
-            gl::FramebufferTexture2D(
-                gl::FRAMEBUFFER,
-                gl::COLOR_ATTACHMENT0,
-                gl::TEXTURE_2D,
-                texture.id,
+            self.gl.framebuffer_texture_2d(
+                glow::FRAMEBUFFER,
+                glow::COLOR_ATTACHMENT0,
+                glow::TEXTURE_2D,
+                Some(texture.id),
                 0,
             );
 
             if rebind_previous {
-                gl::BindFramebuffer(gl::FRAMEBUFFER, previous_id);
+                self.gl.bind_framebuffer(glow::FRAMEBUFFER, previous_id);
                 self.current_framebuffer = previous_id;
             }
         }
@@ -473,74 +420,71 @@ impl GLDevice {
 
     pub fn set_viewport(&mut self, x: i32, y: i32, width: i32, height: i32) {
         unsafe {
-            gl::Viewport(x, y, width, height);
+            self.gl.viewport(x, y, width, height);
         }
     }
 
-    pub fn draw_elements(&mut self, index_buffer: &GLIndexBuffer, count: usize) {
+    pub fn draw_elements(&mut self, index_buffer: &GLIndexBuffer, count: i32) {
         unsafe {
-            self.bind_index_buffer(index_buffer);
+            self.bind_index_buffer(Some(index_buffer));
 
-            gl::DrawElements(
-                gl::TRIANGLES,
-                count as GLsizei,
-                gl::UNSIGNED_INT,
-                ptr::null(),
-            );
+            self.gl
+                .draw_elements(glow::TRIANGLES, count, glow::UNSIGNED_INT, 0);
         }
     }
 
-    pub fn bind_vertex_buffer(&mut self, buffer: &GLVertexBuffer) {
+    pub fn bind_vertex_buffer(&mut self, buffer: Option<&GLVertexBuffer>) {
         unsafe {
-            if self.current_vertex_buffer != buffer.id {
-                gl::BindBuffer(gl::ARRAY_BUFFER, buffer.id);
-                self.current_vertex_buffer = buffer.id;
+            let id = buffer.map(|x| x.id);
+
+            if self.current_vertex_buffer != id {
+                self.gl.bind_buffer(glow::ARRAY_BUFFER, id);
+                self.current_vertex_buffer = id;
             }
         }
     }
 
-    pub fn bind_index_buffer(&mut self, buffer: &GLIndexBuffer) {
+    pub fn bind_index_buffer(&mut self, buffer: Option<&GLIndexBuffer>) {
         unsafe {
-            if self.current_index_buffer != buffer.id {
-                gl::BindBuffer(gl::ELEMENT_ARRAY_BUFFER, buffer.id);
-                self.current_index_buffer = buffer.id;
+            let id = buffer.map(|x| x.id);
+
+            if self.current_index_buffer != id {
+                self.gl.bind_buffer(glow::ELEMENT_ARRAY_BUFFER, id);
+                self.current_index_buffer = id;
             }
         }
     }
 
-    pub fn bind_program(&mut self, program: &GLProgram) {
+    pub fn bind_program(&mut self, program: Option<&GLProgram>) {
         unsafe {
-            if self.current_program != program.id {
-                gl::UseProgram(program.id);
-                self.current_program = program.id;
+            let id = program.map(|x| x.id);
+
+            if self.current_program != id {
+                self.gl.use_program(id);
+                self.current_program = id;
             }
         }
     }
 
-    pub fn bind_texture(&mut self, texture: &GLTexture) {
+    pub fn bind_texture(&mut self, texture: Option<&GLTexture>) {
         unsafe {
-            if self.current_texture != texture.id {
-                gl::ActiveTexture(gl::TEXTURE0);
-                gl::BindTexture(gl::TEXTURE_2D, texture.id);
-                self.current_texture = texture.id;
+            let id = texture.map(|x| x.id);
+
+            if self.current_texture != id {
+                self.gl.active_texture(glow::TEXTURE0);
+                self.gl.bind_texture(glow::TEXTURE_2D, id);
+                self.current_texture = id;
             }
         }
     }
 
-    pub fn bind_framebuffer(&mut self, framebuffer: &GLFramebuffer) {
+    pub fn bind_framebuffer(&mut self, framebuffer: Option<&GLFramebuffer>) {
         unsafe {
-            if self.current_framebuffer != framebuffer.id {
-                gl::BindFramebuffer(gl::FRAMEBUFFER, framebuffer.id);
-                self.current_framebuffer = framebuffer.id;
-            }
-        }
-    }
+            let id = framebuffer.map(|x| x.id);
 
-    pub fn bind_default_framebuffer(&mut self) {
-        unsafe {
-            if self.current_framebuffer != 0 {
-                gl::BindFramebuffer(gl::FRAMEBUFFER, 0);
-                self.current_framebuffer = 0;
+            if self.current_framebuffer != id {
+                self.gl.bind_framebuffer(glow::FRAMEBUFFER, id);
+                self.current_framebuffer = id;
             }
         }
     }
@@ -557,8 +501,11 @@ impl GLDevice {
 impl Drop for GLDevice {
     fn drop(&mut self) {
         unsafe {
-            gl::BindVertexArray(0);
-            gl::DeleteVertexArrays(1, &self.current_vertex_array);
+            self.gl.bind_vertex_array(None);
+
+            if let Some(va) = self.current_vertex_array {
+                self.gl.delete_vertex_array(va);
+            }
         }
     }
 }
@@ -569,11 +516,11 @@ pub enum BufferUsage {
     DynamicDraw,
 }
 
-impl From<BufferUsage> for GLenum {
-    fn from(buffer_usage: BufferUsage) -> GLenum {
+impl From<BufferUsage> for u32 {
+    fn from(buffer_usage: BufferUsage) -> u32 {
         match buffer_usage {
-            BufferUsage::StaticDraw => gl::STATIC_DRAW,
-            BufferUsage::DynamicDraw => gl::DYNAMIC_DRAW,
+            BufferUsage::StaticDraw => glow::STATIC_DRAW,
+            BufferUsage::DynamicDraw => glow::DYNAMIC_DRAW,
         }
     }
 }
@@ -585,12 +532,12 @@ pub enum TextureFormat {
     Red,
 }
 
-impl From<TextureFormat> for GLenum {
-    fn from(texture_format: TextureFormat) -> GLenum {
+impl From<TextureFormat> for u32 {
+    fn from(texture_format: TextureFormat) -> u32 {
         match texture_format {
-            TextureFormat::Rgba => gl::RGBA,
-            TextureFormat::Rgb => gl::RGB,
-            TextureFormat::Red => gl::RED,
+            TextureFormat::Rgba => glow::RGBA,
+            TextureFormat::Rgb => glow::RGB,
+            TextureFormat::Red => glow::RED,
         }
     }
 }
@@ -601,123 +548,84 @@ pub enum FrontFace {
     CounterClockwise,
 }
 
-impl From<FrontFace> for GLenum {
-    fn from(front_face: FrontFace) -> GLenum {
+impl From<FrontFace> for u32 {
+    fn from(front_face: FrontFace) -> u32 {
         match front_face {
-            FrontFace::Clockwise => gl::CW,
-            FrontFace::CounterClockwise => gl::CCW,
+            FrontFace::Clockwise => glow::CW,
+            FrontFace::CounterClockwise => glow::CCW,
         }
     }
 }
 
 #[doc(hidden)]
-impl From<FilterMode> for GLenum {
-    fn from(filter_mode: FilterMode) -> GLenum {
+impl From<FilterMode> for i32 {
+    fn from(filter_mode: FilterMode) -> i32 {
         match filter_mode {
-            FilterMode::Nearest => gl::NEAREST,
-            FilterMode::Linear => gl::LINEAR,
+            FilterMode::Nearest => glow::NEAREST as i32,
+            FilterMode::Linear => glow::LINEAR as i32,
         }
     }
 }
 
+macro_rules! handle_impls {
+    ($name:ty, $delete:ident) => {
+        impl PartialEq for $name {
+            fn eq(&self, other: &$name) -> bool {
+                self.id == other.id
+            }
+        }
+
+        impl Drop for $name {
+            fn drop(&mut self) {
+                unsafe {
+                    self.gl.$delete(self.id);
+                }
+            }
+        }
+    };
+}
+
+#[derive(Debug)]
 pub struct GLVertexBuffer {
-    id: GLuint,
+    gl: Rc<GlContext>,
+
+    id: BufferId,
     count: usize,
     stride: usize,
 }
 
-impl Drop for GLVertexBuffer {
-    fn drop(&mut self) {
-        unsafe {
-            gl::DeleteBuffers(1, &self.id);
-        }
-    }
-}
+handle_impls!(GLVertexBuffer, delete_buffer);
 
 #[derive(Debug)]
 pub struct GLIndexBuffer {
-    id: GLuint,
+    gl: Rc<GlContext>,
+
+    id: BufferId,
     count: usize,
 }
 
-impl Drop for GLIndexBuffer {
-    fn drop(&mut self) {
-        unsafe {
-            gl::DeleteBuffers(1, &self.id);
-        }
-    }
-}
+handle_impls!(GLIndexBuffer, delete_buffer);
 
-#[derive(Debug, PartialEq)]
+#[derive(Debug)]
 pub struct GLProgram {
-    id: GLuint,
+    gl: Rc<GlContext>,
+
+    id: ProgramId,
 }
 
-impl Drop for GLProgram {
-    fn drop(&mut self) {
-        unsafe {
-            gl::DeleteProgram(self.id);
-        }
-    }
-}
-
-mod sealed {
-    use super::*;
-    pub trait UniformValueTypes {}
-    impl UniformValueTypes for i32 {}
-    impl UniformValueTypes for f32 {}
-    impl UniformValueTypes for Mat4 {}
-    impl<'a, T> UniformValueTypes for &'a T where T: UniformValueTypes {}
-}
-
-/// Represents a type that can be passed as a uniform value to a shader.
-///
-/// As the implementation of this trait currently interacts directly with the OpenGL layer,
-/// it's marked as a [sealed trait](https://rust-lang-nursery.github.io/api-guidelines/future-proofing.html#sealed-traits-protect-against-downstream-implementations-c-sealed),
-/// and can't be implemented outside of Tetra. This might change in the future!
-pub trait UniformValue: sealed::UniformValueTypes {
-    #[doc(hidden)]
-    unsafe fn set_uniform(&self, location: GLint);
-}
-
-impl UniformValue for i32 {
-    #[doc(hidden)]
-    unsafe fn set_uniform(&self, location: GLint) {
-        gl::Uniform1i(location, *self);
-    }
-}
-
-impl UniformValue for f32 {
-    #[doc(hidden)]
-    unsafe fn set_uniform(&self, location: GLint) {
-        gl::Uniform1f(location, *self);
-    }
-}
-
-impl UniformValue for Mat4 {
-    #[doc(hidden)]
-    unsafe fn set_uniform(&self, location: GLint) {
-        gl::UniformMatrix4fv(location, 1, gl::FALSE, self.as_slice().as_ptr());
-    }
-}
-
-impl<'a, T> UniformValue for &'a T
-where
-    T: UniformValue,
-{
-    #[doc(hidden)]
-    unsafe fn set_uniform(&self, location: GLint) {
-        (**self).set_uniform(location);
-    }
-}
+handle_impls!(GLProgram, delete_program);
 
 #[derive(Debug)]
 pub struct GLTexture {
-    id: GLuint,
+    gl: Rc<GlContext>,
+
+    id: TextureId,
     width: i32,
     height: i32,
     filter_mode: FilterMode,
 }
+
+handle_impls!(GLTexture, delete_texture);
 
 impl GLTexture {
     pub fn width(&self) -> i32 {
@@ -739,29 +647,71 @@ impl GLTexture {
     }
 }
 
-impl PartialEq for GLTexture {
-    fn eq(&self, other: &GLTexture) -> bool {
-        self.id == other.id
-    }
-}
-
-impl Drop for GLTexture {
-    fn drop(&mut self) {
-        unsafe {
-            gl::DeleteTextures(1, &self.id);
-        }
-    }
-}
-
-#[derive(Debug, PartialEq)]
+#[derive(Debug)]
 pub struct GLFramebuffer {
-    id: GLuint,
+    gl: Rc<GlContext>,
+
+    id: FramebufferId,
 }
 
-impl Drop for GLFramebuffer {
-    fn drop(&mut self) {
-        unsafe {
-            gl::DeleteFramebuffers(1, &self.id);
-        }
+handle_impls!(GLFramebuffer, delete_framebuffer);
+
+mod sealed {
+    use super::*;
+    pub trait UniformValueTypes {}
+    impl UniformValueTypes for i32 {}
+    impl UniformValueTypes for f32 {}
+    impl UniformValueTypes for Mat4 {}
+    impl<'a, T> UniformValueTypes for &'a T where T: UniformValueTypes {}
+}
+
+/// Represents a type that can be passed as a uniform value to a shader.
+///
+/// As the implementation of this trait currently interacts directly with the OpenGL layer,
+/// it's marked as a [sealed trait](https://rust-lang-nursery.github.io/api-guidelines/future-proofing.html#sealed-traits-protect-against-downstream-implementations-c-sealed),
+/// and can't be implemented outside of Tetra. This might change in the future!
+pub trait UniformValue: sealed::UniformValueTypes {
+    #[doc(hidden)]
+    unsafe fn set_uniform(&self, program: &GLProgram, location: Option<u32>);
+}
+
+impl UniformValue for i32 {
+    #[doc(hidden)]
+    unsafe fn set_uniform(&self, program: &GLProgram, location: Option<u32>) {
+        program.gl.uniform_1_i32(location, *self);
+    }
+}
+
+impl UniformValue for f32 {
+    #[doc(hidden)]
+    unsafe fn set_uniform(&self, program: &GLProgram, location: Option<u32>) {
+        program.gl.uniform_1_f32(location, *self);
+    }
+}
+
+impl UniformValue for Mat4 {
+    #[doc(hidden)]
+    unsafe fn set_uniform(&self, program: &GLProgram, location: Option<u32>) {
+        let slice = self.as_slice();
+
+        // I don't think there's a scenario where this wouldn't be true,
+        // but better safe than sorry.
+        assert_eq!(slice.len(), 16);
+
+        let array_ref = slice.as_ptr() as *const [f32; 16];
+
+        program
+            .gl
+            .uniform_matrix_4_f32_slice(location, false, &*array_ref);
+    }
+}
+
+impl<'a, T> UniformValue for &'a T
+where
+    T: UniformValue,
+{
+    #[doc(hidden)]
+    unsafe fn set_uniform(&self, program: &GLProgram, location: Option<u32>) {
+        (**self).set_uniform(program, location);
     }
 }